+++ include/pog/parser.h
@@ -3,6 +3,23 @@
 #include <deque>
 #include <unordered_map>
 
+#include <fmt/format.h>
+
+#ifdef POG_DEBUG
+#define POG_DEBUG_PARSER 1
+#endif
+
+#ifdef POG_DEBUG_PARSER
+template <typename... Args>
+void debug_parser(Args&&... args)
+{
+	fmt::print(stderr, "[parser] {}\n", fmt::format(std::forward<Args>(args)...));
+}
+#else
+template <typename... Args>
+void debug_parser(Args&&...) {}
+#endif
+
 #include <pog/action.h>
 #include <pog/automaton.h>
 #include <pog/errors.h>
@@ -99,6 +116,11 @@ public:
 		_grammar.set_start_symbol(_grammar.add_symbol(SymbolKind::Nonterminal, name));
 	}
 
+	void enter_tokenizer_state(const std::string& state_name)
+	{
+		_tokenizer.enter_state(state_name);
+	}
+
 	void push_input_stream(std::istream& input)
 	{
 		_tokenizer.push_input_stream(input);
@@ -111,6 +133,8 @@ public:
 
 	std::optional<ValueT> parse(std::istream& input)
 	{
+		_tokenizer.enter_state(std::string{decltype(_tokenizer)::DefaultState});
+
 		std::optional<TokenMatchType> token;
 		_tokenizer.push_input_stream(input);
 
@@ -129,7 +153,13 @@ public:
 					auto expected_symbols = _parsing_table.get_expected_symbols_from_state(_automaton.get_state(stack.back().first));
 					throw SyntaxError(expected_symbols);
 				}
+
+				debug_parser("Tokenizer returned new token with symbol \'{}\'", token.value().symbol->get_name());
 			}
+			else
+				debug_parser("Reusing old token with symbol \'{}\'", token.value().symbol->get_name());
+
+			debug_parser("Top of the stack is state {}", stack.back().first);
 
 			const auto* next_symbol = token.value().symbol;
 			auto maybe_action = _parsing_table.get_action(_automaton.get_state(stack.back().first), next_symbol);
@@ -144,13 +174,14 @@ public:
 			if (std::holds_alternative<ReduceActionType>(action))
 			{
 				const auto& reduce = std::get<ReduceActionType>(action);
+				debug_parser("Reducing by rule \'{}\'", reduce.rule->to_string());
 
 				// Each symbol on right-hand side of the rule should have record on the stack
 				// We'll pop them out and put them in reverse order so user have them available
 				// left-to-right and not right-to-left.
 				std::vector<ValueT> action_arg;
-				action_arg.reserve(reduce.rule->get_rhs().size());
-				assert(stack.size() >= reduce.rule->get_rhs().size() && "Stack is too small");
+				action_arg.reserve(reduce.rule->get_number_of_required_arguments_for_action());
+				assert(stack.size() >= action_arg.capacity() && "Stack is too small");
 
 				for (std::size_t i = 0; i < action_arg.capacity(); ++i)
 				{
@@ -158,31 +189,55 @@ public:
 					// We need to do this in order to perform move together with value_or()
 					// See: https://en.cppreference.com/w/cpp/utility/optional/value_or
 					// std::move(*this) is performed only when value_or() is called from r-value
-					action_arg.insert(action_arg.begin(), std::move(stack.back().second).value_or(ValueT{}));
-					stack.pop_back();
+					//
+					// Also do not pop from stack here because midrule actions can still return us arguments back
+					action_arg.insert(action_arg.begin(), std::move(stack[stack.size() - i - 1].second).value_or(ValueT{}));
 				}
 
 				// What left on the stack now determines what state we get into now
-				auto maybe_next_state = _parsing_table.get_transition(_automaton.get_state(stack.back().first), reduce.rule->get_lhs());
+				// We use size of RHS to determine stack top because midrule actions might have only borrowed something from stack so the
+				// real stack top is not the actual top. Midrule actions have 0 RHS size even though they borrow items. Other rules
+				// have same size of RHS and what they take out of stack.
+				auto maybe_next_state = _parsing_table.get_transition(_automaton.get_state(stack[stack.size() - reduce.rule->get_rhs().size() - 1].first), reduce.rule->get_lhs());
 				if (!maybe_next_state)
 				{
 					assert(false && "Reduction happened but corresponding GOTO table record is empty");
 					return std::nullopt;
 				}
 
+				auto action_result = reduce.rule->has_action() ? reduce.rule->perform_action(std::move(action_arg)) : ValueT{};
+
+				// Midrule actions only borrowed arguments and it is returning them back
+				if (reduce.rule->is_midrule())
+				{
+					for (std::size_t i = 0; i < action_arg.size(); ++i)
+						stack[stack.size() - i - 1].second = std::move(action_arg[action_arg.size() - i - 1]);
+				}
+				// Non-midrule actions actually consumed those arguments so pop them out
+				else
+				{
+					for (std::size_t i = 0; i < action_arg.size(); ++i)
+						stack.pop_back();
+				}
+
+				debug_parser("Pushing state {}", maybe_next_state.value()->get_index());
+
 				stack.emplace_back(
 					maybe_next_state.value()->get_index(),
-					reduce.rule->has_action() ? reduce.rule->perform_action(std::move(action_arg)) : ValueT{}
+					std::move(action_result)
 				);
 			}
 			else if (std::holds_alternative<ShiftActionType>(action))
 			{
+				const auto& shift = std::get<ShiftActionType>(action);
+				debug_parser("Shifting state {}", shift.state->get_index());
+
 				// Notice how std::move() is only around optional itself and not the whole expressions
 				// We need to do this in order to perform move together with value()
 				// See: https://en.cppreference.com/w/cpp/utility/optional/value
 				// Return by rvalue is performed only when value() is called from r-value
 				stack.emplace_back(
-					std::get<ShiftActionType>(action).state->get_index(),
+					shift.state->get_index(),
 					std::move(token).value().value
 				);
 
@@ -191,6 +246,7 @@ public:
 			}
 			else if (std::holds_alternative<Accept>(action))
 			{
+				debug_parser("Accept");
 				// Notice how std::move() is only around optional itself and not the whole expressions
 				// We need to do this in order to perform move together with value()
 				// See: https://en.cppreference.com/w/cpp/utility/optional/value