+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# dev
+
+* Fixed calulcation of includes and lookback relations when there are more instances of the same symbol inspected
+* Added option to define `POG_DEBUG` to print debugging messages from parser adn tokenizer
+
+# v0.3.0 (2019-09-22)
+
+* Midrule actions and all preceding symbols are now accessible from later actions in that rule
+* Explicit switching of tokenizer state with `enter_tokenizer_state` method
+* Implicit end of input token now has modifiable states in which it is active
+
 # v0.2.1 (2019-09-13)
 
 * Added option `POG_PIC` to build position-independent code